+++ css/colors.php
@@ -1,0 +1,36 @@
+<?php
+    header("Content-type: text/css");
+    require("../php/loadColors.php");
+?>
+
+.c1_cl {
+    color: <?php echo $C1 ?> !important;
+}
+
+.c1_bg {
+    background-color: <?php echo $C1 ?> !important;
+}
+
+.c2_cl {
+    color: <?php echo $C2 ?> !important;
+}
+
+.c2_bg {
+    background-color: <?php echo $C2 ?> !important;
+}
+
+.c3_cl {
+    color: <?php echo $C3 ?> !important;
+}
+
+.c3_bg {
+    background-color: <?php echo $C3 ?> !important;
+}
+
+.c4_cl {
+    color: <?php echo $C4 ?> !important;
+}
+
+.c4_bg {
+    background-color: <?php echo $C4 ?> !important;
+}