+++ index.php
@@ -15,12 +15,15 @@
 <!DOCTYPE html>
 <html>
     <head>
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
         <link rel="icon" href="favicon.ico" type="image/gif" sizes="16x16">
         <link rel="stylesheet" type="text/css" href="css/main.css">
         <link rel="stylesheet" type="text/css" href="css/index/editing.css">
         <link rel="stylesheet" type="text/css" href="css/navbar.css">
         <link rel="stylesheet" type="text/css" href="css/index/container.css">
+        <link rel="stylesheet" type="text/css" href="css/colors.php">
+        <link href="https://fonts.googleapis.com/icon?family=Material+Icons+Outlined" rel="stylesheet">
         <script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
         <title>Startpage</title>
         <script>
@@ -46,7 +49,7 @@
                             $(".linkA").each(function() {
                                 var linkName = $(this).html();
                                 var linkHref = $(this).attr("href");
-                                var searchLink = '<a class="searchLink" href="'+linkHref+'">'+linkName+'</a>';
+                                var searchLink = '<a class="searchLink c3_cl" href="'+linkHref+'">'+linkName+'</a>';
 
                                 if(linkName.toLowerCase().includes(input) || linkHref.toLowerCase().includes(input)) {
                                     $(searchLink).appendTo(".searchResults");
@@ -130,15 +133,15 @@
             })
         </script>
     </head>
-    <body>
-        <nav class="topnav">
-            <input type="search" class="searchBar" placeholder="Search">
-            <ul class="searchResults" style="display: none;">
+    <body class="c1_bg">
+        <nav class="topnav c3_bg">
+            <input type="search" class="searchBar c4_bg c3_cl" placeholder="Search">
+            <ul class="searchResults c4_bg c3_cl" style="display: none;">
             </ul>
             <div class="navIconContainer">
                 <input class="editStatus" value="false" disabled hidden>
-                <span class="editBtn btn"><img class="navIcon editIcon" src="_assets/edit-24px.svg" alt="editIcon"></span>
-                <a href="pages/settings.php"><img class="navIcon settingsIcon" src="_assets/settings-24px.svg" alt="settingsIcon"></a>
+                <span class="editBtn btn"><i class="navIcon editIcon material-icons-outlined c2_cl">create</i></span>
+                <a href="pages/settings.php" style="text-decoration: none;"><i class="navIcon settingsIcon material-icons-outlined c2_cl">settings</i></a>
             </div>
         </nav>
         <div class="contentWrapper">